Added the default sound onto the list. You cannot delete the defaults from the list. Also added the translations.

+++ demo/src/com/novoda/moo/demo/FileMgr.java
@@ -1,6 +1,10 @@
 package com.novoda.moo.demo;
 
 import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.ArrayList;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
@@ -9,6 +13,7 @@ import android.content.ContentResolver;
 import android.content.ContentValues;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.res.Resources.NotFoundException;
 import android.media.RingtoneManager;
 import android.net.Uri;
 import android.os.Bundle;
@@ -25,8 +30,10 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ContextMenu.ContextMenuInfo;
 import android.widget.AdapterView;
+import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ListView;
+import android.widget.TextView;
 import android.widget.AdapterView.OnItemClickListener;
 
 import com.novoda.moo.demo.os.FileSys;
@@ -44,27 +51,39 @@ public class FileMgr extends ListActivity {
 	public static String			RINGTONES_DIR;
 	public static String			AUDIO_FILES_DIR;
 	public static String			TMP_AUDIO_DIR;
+	private static String			DEFAULT_AUDIO_DIR;
+	private Button					banner;
 
 	@Override
 	public void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.filemgr_demo);
-
+		
 		if (android.os.Environment.getExternalStorageState().equals(android.os.Environment.MEDIA_MOUNTED)) {
 			Log.d(TAG, "SDCard installed - Checking for sound directories");
 			AUDIO_FILES_DIR = Environment.getExternalStorageDirectory() + this.getString(R.string.dir_created_sounds);
 			RINGTONES_DIR = Environment.getExternalStorageDirectory() + this.getString(R.string.dir_created_ringtones);
 			TMP_AUDIO_DIR = Environment.getExternalStorageDirectory() + this.getString(R.string.dir_tmp);
+			DEFAULT_AUDIO_DIR = Environment.getExternalStorageDirectory() + this.getString(R.string.dir_default_audio);
 
 			new File(AUDIO_FILES_DIR).mkdirs();
 			new File(RINGTONES_DIR).mkdirs();
 			new File(TMP_AUDIO_DIR).mkdirs();
+			new File(DEFAULT_AUDIO_DIR).mkdirs();
 
 			initFileListing();
 		} else {
 			Log.i(TAG, "No SDCard installed, no sound files can be saved.");
 			showDialog(R.layout.dialog_sdcard_warning);
 		}
+
+		banner = (Button) findViewById(R.id.mooAdvert);
+		banner.setOnClickListener(new View.OnClickListener() {
+			public void onClick(View v) {
+				Intent viewIntent = new Intent("android.intent.action.VIEW", Uri.parse("http://www.novoda.com"));
+				startActivity(viewIntent);
+			}
+		});
 
 	}
 
@@ -76,23 +95,34 @@ public class FileMgr extends ListActivity {
 		Log.i(TAG, "Name of file=" + sCurrFileName);
 		boolean success = false;
 
+		
 		switch (item.getItemId()) {
 			case R.id.delete:
-				Log.d(TAG, "Deleting file [" + sCurrFileName + "]");
-				success = (new File(AUDIO_FILES_DIR + sCurrFileName)).delete();
-
-				if (success) {
-					// TODO: DELETE entry in Media DB via URI
-					Log.i(TAG, "File [" + sCurrFileName + "] was deleted");
-					refreshFileListAdapter();
-				} else {
-					Log.e(TAG, "File  [" + sCurrFileName + "] was not deleted");
+				
+				if(new File(DEFAULT_AUDIO_DIR + sCurrFileName).exists()){
+					showDialog(R.layout.dialog_not_on_default_file);
+				}else{
+					Log.d(TAG, "Deleting file [" + sCurrFileName + "]");
+					success = (new File(AUDIO_FILES_DIR + sCurrFileName)).delete();
+					
+					if (success) {
+						// TODO: DELETE entry in Media DB via URI
+						Log.i(TAG, "File [" + sCurrFileName + "] was deleted");
+						refreshFileListAdapter();
+					} else {
+						Log.e(TAG, "File  [" + sCurrFileName + "] was not deleted");
+					}
 				}
 				break;
 
 			case R.id.rename:
-				Log.d(TAG, "Renaming file [" + sCurrFileName + "]");
-				showDialog(R.layout.dialog_rename);
+				
+				if(new File(DEFAULT_AUDIO_DIR + sCurrFileName).exists()){
+					showDialog(R.layout.dialog_not_on_default_file);
+				}else{
+					Log.d(TAG, "Renaming file [" + sCurrFileName + "]");
+					showDialog(R.layout.dialog_rename);
+				}
 				break;
 
 			case R.id.email:
@@ -119,6 +149,7 @@ public class FileMgr extends ListActivity {
 
 		return true;
 	}
+
 
 	@Override
 	protected Dialog onCreateDialog(int id) {
@@ -182,6 +213,18 @@ public class FileMgr extends ListActivity {
 							public void onClick(DialogInterface dialog, int whichButton) {
 							}
 						}).create();
+				
+			case R.layout.dialog_not_on_default_file:
+				LayoutInflater factory4 = LayoutInflater.from(this);
+				final View textEntryView4 = factory4.inflate(R.layout.dialog_not_on_default_file, null);
+
+				return new AlertDialog.Builder(FileMgr.this)
+					.setIcon(android.R.drawable.ic_dialog_info)
+					.setView(textEntryView4)
+					.setPositiveButton(R.string.dialog_ok, new DialogInterface.OnClickListener() {
+						public void onClick(DialogInterface dialog, int whichButton) {
+						}
+					}).create();
 		}
 
 		return null;
@@ -194,7 +237,6 @@ public class FileMgr extends ListActivity {
 	}
 
 	public boolean onOptionsItemSelected(MenuItem item) {
-
 		switch (item.getItemId()) {
 			case R.id.help:
 				Log.v(TAG, "User selected Help");
@@ -233,17 +275,55 @@ public class FileMgr extends ListActivity {
 
 				Intent intent = new Intent();
 				intent.putExtra(Constants.PICKED_AUDIO_FILE_POSITION, position);
+				Log.i(TAG, "Row title : " + ((TextView) view.findViewById(R.id.row_title)).getText());
+				intent.putExtra(Constants.PICKED_AUDIO_FILE_NAME, view.findViewById(R.id.row_title).toString());
 				setResult(RESULT_OK, intent);
 				finish();
 			}
 		});
-		
-		
+
 	}
 
 	private void refreshFileListAdapter() {
 		mFileListAdapter = new AudioFileListAdapter(this, FileSys.listFilesInDir(AUDIO_FILES_DIR));
+		mFileListAdapter = addSoundDefaults(mFileListAdapter);
 		this.setListAdapter(mFileListAdapter);
+	}
+	
+	private AudioFileListAdapter addSoundDefaults(AudioFileListAdapter fileListAdapter) {
+		String defaultSoundFile = Environment.getExternalStorageDirectory() +  this.getString(R.string.dir_default_audio) + "carlcow.wav";
+		File defaultSoundFile1 = new File(defaultSoundFile);
+		
+		if(!defaultSoundFile1.exists()){
+			writeResToFile(this.getResources().openRawResource(R.raw.carlthecow), defaultSoundFile);
+		}
+		
+		ArrayList<File> listFilesInDir = FileSys.listFilesInDir(Environment.getExternalStorageDirectory() +  this.getString(R.string.dir_default_audio));
+		fileListAdapter.files.addAll(listFilesInDir);
+	    return fileListAdapter;
+	}
+
+	private File writeResToFile(InputStream ins, String fileName) {
+		int size;
+		File file = null;
+		
+		try {
+			Runtime.getRuntime().exec("chmod 666 " + Environment.getExternalStorageDirectory() +  this.getResources().getString(R.string.dir_default_audio));
+			file = new File(fileName);
+			size = ins.available();
+            byte[] buffer = new byte[size];
+            ins.read(buffer);
+            ins.close();
+            FileOutputStream fos = new FileOutputStream(file);
+            fos.write(buffer);
+            fos.close();
+        } catch (NotFoundException e1) {
+            Log.e(TAG, "Problem with creating default files on external storage", e1);
+        } catch (IOException e2) {
+            Log.e(TAG, "Problem with creating default files on external storage", e2);
+        }
+        
+        return file;
 	}
 
 	/***
